neatserpupdate

+++ database/migrations/2026_07_11_092339_create_materials_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->string('nama');
             $table->string('kategori');
             $table->string('satuan');
-            $table->decimal('stok', 10, 2)->default(0);
-            $table->decimal('stok_minimum', 10, 2)->default(0);
+            $table->integer('stok')->default(0);
+            $table->integer('stok_minimum')->default(0);
             $table->decimal('harga', 15, 2)->default(0);
             $table->timestamps();
         });
